requirements.txt
Django
django-nested-admin